Use visitor pattern for AST
Instead of adding methods to the nodes we can make use of visitor.
This makes the nodes accept a visitor.
The visitor generic such that it can return whatever it needs.
There are currently implemented two visitors:
PrettyPrintingVisitor and TreePrintingVisitor.
The pretty printing visitor prints the AST as source code.
The tree printing visitor prints the AST itself.