Add Point
Make a TernaryNode abstract node and make DeclarationNode use it.
Make move take a point instead of three numbers.

A solution should be made in the future regarding type checking
combinations of point and number and such.

A similar solution could be used for the interactions of these
types through operators.