Implement function calls in code generator
This is a very limited and hacky approach.
There are only one symbol table and thus only one scope.
The formal parameters for a function is written into the global
symbol table with the actual parameter values.
The function body statement list is then visited.
So nested function calls are hopelessly broken.

But this is just the first step to a complete implementation.